Check for error in schema
Sometimes the API returns an error code that was getting interpreted as a schema rather than an error. This checks for the existence of an error field in the returned code and gives the error message when that happens.

Changed the downloading to be a fallback situation if the parsing of a file fails rather than an either-or thing.