Use FileSaver.js, progess bar, moved link to cog
The June 4, 2016, release consolidated buttons in the People page. This moves the Access Report to the cog with the rest of the menu items.

This release adds a progress bar.

This script now uses the RequireJS library to include the FileSaver.js library. That makes saving files easier and better tested with different OSes