Changed primary key to unique key
Changed primary key to unique key seems to be necessary to keep duplicates from creeping in. There is an ALTER TABLE IGNORE command to add the index after the table has been created, but it is deprecated as of MySQL 5.7.4. After 5.7.4, removing the duplicates after they are already in there involves creating another table. http://www.tocker.ca/2013/11/06/the-future-of-alter-ignore-table-syntax.html

It is HIGHLY RECOMMENDED that you sort -u your data before importing it. It will speed up the process tremendously.