Added type="button"
The default type for a button is submit and when the user would hit enter to change a point value for a question, it would execute the first button on the page of type submit. Since the buttons to run a method didn't have a type, they were getting executed when they appeared first on the page instead of the update scores button.