Safari and Chrome fixes for autoAdvance
Dumped the use of .next in favor of quizwiz_next just in case it conflicted with Canvas' use of .next
Changed the &> to just > and made it narrower.
When clicking on a button in Firefox, the button was returned as the target. However, in Safari, the item within the button was returned as the target. This was causing the autoAdvance to not work.