Fixes for feedback left with questions
Feedback left with questions are marked with a class of quiz_comment, but they don't have a textarea comment box. These questions were throwing an error.
Once that was fixed, it became apparent that I was finding the first case of quiz_comment instead of the last because I wasn't breaking the for loop once it was found.