Fix advance to next user, cross-domain content
The element being watched for changes to the grade stopped being modified in the way I was checking and the advance on update and advance on saved rubric stopped working.  I changed the element I watch to the count of students who have been graded, which Canvas rescans and updates after every update to the grade, even if nothing has changed. Thanks to Dallas Hulsey for finding this.

In the debugging, we also noticed that there was a cross-domain iframe error, presumably from Turnitin. Added a separate check to help suppress the error.