Eliminate requirejs requirement
Canvas had a note that they're not using requirejs anymore and that people should take care of loading their own functions. Rather than continue to use requirejs and load it myself, I now load the FileSaver directly.
Added a headingNoSpaces configuration variable that will remove spaces from the headings when set to true. I was told that R has problems with spaces in the variable names.
Added a showViewStudent configuration variable that will remove any asset_class_name = 'student_enrollment' from the data before downloading unless it is true. Canvas has started marking viewing another student's profile page as a participation and the spreadsheet was confusing faculty who wondered why their student names were showing up as titles. Removing the data is the default, but you can set showViewStudent = true to include the data in the download.
Moved the tabindex=-1 to the anchor tag rather than the list tag when creating the menu item so that it matches what Canvas does for the other items