Change Boolean to ENUM('false','true')
Canvas data uses true and false for boolean fields, while MySQL uses 1 and 0. This meant that all of the boolean data was getting lost when the file was loaded.
You can convert the data, but there isn't an easy way to know which fields to convert since 'BOOLEAN' is an alias for TINYINT and the information_schema.column table lists it as TINYINT instead of BOOLEAN. This means that the import.sh file gets updated with custom import commands for each new schema or we switch to ENUM('false','true') for BOOLEAN types. This means you'll need to be explicit when working with boolean fields, but the information wasn't there at all before, so it shouldn't be a big issue.