moved to People page, fixed saveAs
Eli Grey's saveAs moved locations yesterday. However, the version as I write this (1.3.5) does not work with Tampermonkey and so I'm grabbing an older version from a CDN rather than the latest version.

Moved the link from the Account / Courses page to the Account / People page. It now appears in the top section with View User Groups and Manage Profile Pictures. Arguably, that might have been a better location in the first place. The reason for this move is two-fold. 
* Canvas' new course search doesn't have the places needed to add the button and although this was under a feature flag, it might eventually disappear.
* Canvas moved the Analytics link off the main courses page to the account navigation menu on December 9, 2017. The script was looking for the link to know where to add the button and it hasn't been working since then.