Use the correct month
JavaScript months are 0-based where Jan = 0 and Dec = 11. I was using the human scheme where Jan = 1 and Dec = 12. I caught the mistake when Jan 31 was showing up as later than Feb 2. It was converting Date(2018,1,31) to March 3 and Date(2018,2,2) to March 2.
Thanks go to the student who hadn't logged into the course in 3 days for helping me find the bug.