Major rewrite
This removes the jQuery dependency and uses the window.fetch() command. This is supported by modern browsers, but not by Internet Explorer. In other words, this will not work with Internet Explorer.
This changes the breadcrumb and its links
This changes the document title that shows up on the printed page
It now checks for the student grades table and works backwards to find a heading, rather than just looking for "Courses" in the heading. This keeps it from changing "Courses I'm Teaching" and works better with other languages.