Fixed submission comments autoadvance
Canvas renamed the selector for submission comments from speedgrader_comment_textarea to speed_grader_comment_textarea. This happened with the September 15, 2018, release as part of adding submission comments to a PDF file for download.

For the time being, I'm checking for both speed_grader_comment_textarea and speedgrader_comment_textarea so those with self-hosted systems that are not up-to-date continue to function.