Don't rely on Canvas' jQuery
The user script runs before jQuery is available from Canvas so $ was not available.
The script no requires both jQuery and the Tablesorter plugin.
The problem is that by the time that the tablesorter is needed, Canvas' jQuery has overwritten the local version loaded by the userscript manager. We save the local jQuery as jq and then refer to it instead of $.

Made changes so that this will work from the custom global JavaScript. If the tablesorter plugin is not available, then it loads it dynamically. Canvas' jQuery is very likely available by the time that it does this, so if it is, then tablesorter gets loaded into Canvas' jQuery instead of the local copy. We then use Canvas' copy of jQuery instead of the previously saved local version.