Fetch quiz questions without a submission
About 2.5 years ago, Canvas stopped delivering the author's question name when fetching questions for a specific submission. Instead, it started returning Question 1, Question 2, etc., which is worthless since that is already there.

Forcing the questions for an entire quiz can take a while, especially if there are a lot of questions (think question groups), so there was a noticeable delay in processing. To work around this, I set the cache mode to force-cache, which first looks in the browser for a copy of the request and response and uses it without even checking for an up-to-date version from Canvas. It does this very naively and so if you do happen to update the question names after the students have taken the quiz and stale data is in your browser's cache, you will need to clear the cache with Ctrl+Shift+Del.

This update makes several changes
* Switch to download the full set of questions rather than just those given on the quiz submission you are looking at.
* Removed the dependency on jQuery, instead using the browser's fetch() function.
* Force the cache to force-cache mode.
* Don't append the question name if it is the same as the existing text.