Bypass sentry so eventlisteners work in an iframe
Canvas has added Sentry, which intercepts all addEventListener calls. However, it doesn't work right with iframes and so the quiz enhancements stopped working.
This finds the original addEventListener for EventTarget and calls it instead of using the Sentry version.
This is a minor update of the existing code that just fixes the event listeners. This only looks like a big update because the code formatter I used made a lot of changes.
I'm working on a broader rewrite that needs a lot of testing, but wanted to get something out for people to use quickly.