Bypass Canvas' use of sentry
Canvas started using Sentry, which grabs all addEventListener methods and runs it through the Sentry system.
This broke event listeners on iframes, which is where Classic Quizzes in SpeedGrader is loaded.
This patch checks for sentry and if it's there, grabs the original addEventListener and uses it instead.
There is a bigger rewrite in the works, but it needs more testing. This is a small change (other than formatting) designed to get things out to people quickly.