Fix missing data due to last page
The enrollment request is expensive and Canvas doesn't always return the last page link header right away.
That meant that concurrent requests were not being made and they were being processed sequentially.
The last page link header did finally appear (for me it was at the next to last page) and it failed to the last page.
I noticed that when the sequential loading happens for large classes that the progress bar is really inaccurate since it doesn't know how many students there are. That will have to be addressed in a different patch.
I also made some changes to the metadata based on lint messages that Tampermonkey was giving.