Update to work with Canvas' solution
Canvas added their own sorting in a very non-efficient, reload the entire page, way. That broke this script.
Canvas has updated their jQuery to a more modern version.
Removed the check (which appears to have been wrong) for jQuery.
Clicking on the header will use Canvas sort. Click to the right of the header in the empty space to use this script.